Logo update and Navigation
Logo has been updated to a simple letter 'H', removed the border and background colour.
Increased the font family, size and weight for the Logo.

Added a Navigation button to the home screen that is will expand and collapse
correspondingly showing and hiding the menu.
Currently there is only a placeholder menu, which will be updated when new screens are added.